third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/dialog-return-value-expected.txt

Tests that dialog.returnValue is settable and returns the last value set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dialog.returnValue is ''
PASS dialog.returnValue is 'Setting value directly'
PASS dialog.returnValue is 'null'
PASS dialog.returnValue is ''
PASS dialog.returnValue is '7'
PASS dialog.returnValue is 'Return value set from close()'
PASS dialog.returnValue is ''
PASS dialog.returnValue is 'null'
PASS dialog.returnValue is 'Should not change because no argument to close()'
PASS dialog.close('blah') threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS dialog.returnValue is 'Should not change because close() is invalid'
PASS successfullyParsed is true

TEST COMPLETE
